% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{insertDivAddin}
\alias{insertDivAddin}
\title{Insert div algin center image}
\usage{
insertDivAddin()
}
\value{
html syntex div center and image
}
\description{
This will generate the html syntax for align center and insert image
}
